feat: #39 BloomState DesignKit 이동
