+++ Projects/Features/Blooming/BloomingFeatureInterface/Sources/BloomingUpdateView.swift
@@ -8,7 +8,6 @@
 
 import SwiftUI
 import DesignKit
-import Utility
 import ComposableArchitecture
 
 public struct BloomingUpdateView: View {
